requirements.txt
streamlit
rembg
opencv-python-headless
numpy
Pillow
streamlit-drawable-canvas